Fix crashes on video playback.
Change I8ee9162536da5711e383dd1d89dca4fa1da107be missed correct variable
initializations at one of two places, which led to a null pointer
dereference. This patch fixes that.
It also updates the variable that stores the original allocated length
to be of the same type as used in the OMX buffer header.

Change-Id: I15b083da8565ec3e8769d18ba1f19d3b12f54cb1